Tricky color fix.
Previous check sum where unable to detect collisions if the colors where
able to mix, like 1 and 6 or'ed together will create the value 7. in
that case, the check sum calculation wouldnt detect the collision.

Instead of just calculating the sum from all elements, we calculate the
number of valued elements.